+++ .github/workflows/ci-delta.yml
@@ -37,8 +37,8 @@ jobs:
         run: sbt -Dsbt.color=always -Dsbt.supershell=false "project delta" clean scalafmtCheck test:scalafmtCheck scalafmtSbtCheck scapegoat
       - name: Tests
         run: |
-          echo "127.0.0.1 bucket.my-domain.com" | sudo tee -a /etc/hosts
-          echo "127.0.0.1 bucket2.my-domain.com" | sudo tee -a /etc/hosts
+          grep -qF 'bucket.my-domain.com' /etc/hosts || echo "127.0.0.1 bucket.my-domain.com" | sudo tee -a /etc/hosts
+          grep -qF 'bucket2.my-domain.com' /etc/hosts || echo "127.0.0.1 bucket2.my-domain.com" | sudo tee -a /etc/hosts
           sbt -Dsbt.color=always -Dsbt.supershell=false "project delta" clean coverage test coverageReport coverageAggregate
   publish:
     if: github.event_name == 'push'
